second commit

+++ src/main/java/com/ltqtest/springbootquickstart/repository/UserRepository.java
@@ -10,5 +10,5 @@ import java.util.Optional;
 public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByUsername(String username);
     boolean existsByUsername(String username);
-    Optional<User> findByUserId(Long userId);
+    Optional<User> findByUserId(Integer userId);
 }
